Moved Checkout components into their own folder.
